refactor: extract logging render edge into logging module
